+++ services/synccommitteemessenger/standard/service_test.go
@@ -15,6 +15,9 @@ package standard_test
 
 import (
 	"context"
+	"github.com/attestantio/go-eth2-client/spec/phase0"
+	"github.com/attestantio/vouch/services/synccommitteemessenger"
+	"github.com/attestantio/vouch/testutil"
 	"testing"
 	"time"
 
@@ -278,6 +281,175 @@ func TestService(t *testing.T) {
 				}
 			} else {
 				require.NoError(t, err)
+			}
+		})
+	}
+}
+
+func TestServiceMessage(t *testing.T) {
+	ctx := context.Background()
+
+	genesisTime := time.Now()
+	genesisProvider := mock.NewGenesisProvider(genesisTime)
+	specProvider := mock.NewSpecProvider()
+	chainTime, err := standardchaintime.New(ctx,
+		standardchaintime.WithLogLevel(zerolog.Disabled),
+		standardchaintime.WithGenesisProvider(genesisProvider),
+		standardchaintime.WithSpecProvider(specProvider),
+	)
+	require.NoError(t, err)
+
+	nullSubmitter, err := nullsubmitter.New(ctx)
+	require.NoError(t, err)
+	mockETH2Client, err := mocketh2client.New(ctx)
+	require.NoError(t, err)
+
+	mockValidatingAccountsProvider := mockaccountmanager.NewValidatingAccountsProvider()
+	mockSyncCommitteeAggregator := mocksynccommitteeaggregator.New()
+	mockSigner := mocksigner.New()
+
+	epoch := phase0.Epoch(100)
+	slot := phase0.Slot(epoch * 32)
+	validatorIndexToCommitteeIndices := testutil.CreateValidatorIndexToCommitteeIndicesTestData()
+	validatorIndices := make([]phase0.ValidatorIndex, len(validatorIndexToCommitteeIndices))
+	for validatorIndex := range validatorIndexToCommitteeIndices {
+		validatorIndices[validatorIndex] = validatorIndex
+	}
+
+	duty := synccommitteemessenger.NewDuty(slot, validatorIndexToCommitteeIndices)
+	accounts, err := testutil.CreateTestWalletAndAccounts(validatorIndices, "0x25295f0d1d592a90b333e26e85149708208e9f8e8bc18f6c77bd62f8ad7a6866")
+	require.NoError(t, err)
+
+	for validIndex, testAccount := range accounts {
+		duty.SetAccount(validIndex, testAccount)
+		mockValidatingAccountsProvider.AddAccount(validIndex, testAccount)
+	}
+
+	tests := []struct {
+		name     string
+		params   []standard.Parameter
+		logEntry string
+	}{
+		{
+			name: "SetsLastSyncHead",
+			params: []standard.Parameter{
+				standard.WithLogLevel(zerolog.Disabled),
+				standard.WithProcessConcurrency(1),
+				standard.WithMonitor(nullmetrics.New(ctx)),
+				standard.WithChainTimeService(chainTime),
+				standard.WithSyncCommitteeAggregator(mockSyncCommitteeAggregator),
+				standard.WithSpecProvider(specProvider),
+				standard.WithBeaconBlockRootProvider(mockETH2Client),
+				standard.WithSyncCommitteeMessagesSubmitter(nullSubmitter),
+				standard.WithValidatingAccountsProvider(mockValidatingAccountsProvider),
+				standard.WithSyncCommitteeRootSigner(mockSigner),
+				standard.WithSyncCommitteeSelectionSigner(mockSigner),
+				standard.WithSyncCommitteeSubscriptionsSubmitter(nullSubmitter),
+			},
+		},
+	}
+
+	for _, test := range tests {
+		t.Run(test.name, func(t *testing.T) {
+			messageService, err := standard.New(ctx, test.params...)
+			require.NoError(t, err)
+			messages, msgErr := messageService.Message(ctx, duty)
+			require.NoError(t, msgErr)
+			require.NotEmpty(t, messages)
+			lastReported, found := messageService.GetDataUsedForSlot(slot)
+			require.Equal(t, true, found)
+			beaconRootResponse, _ := mockETH2Client.BeaconBlockRoot(context.Background(), nil)
+			expectedRoot := *beaconRootResponse.Data
+			require.Equal(t, expectedRoot, lastReported.Root)
+			require.Equal(t, validatorIndexToCommitteeIndices, lastReported.ValidatorToCommitteeIndex)
+		})
+	}
+}
+
+func TestServiceRemoveHistoricDataUsedForSlotValidation(t *testing.T) {
+	ctx := context.Background()
+
+	genesisTime := time.Now()
+	genesisProvider := mock.NewGenesisProvider(genesisTime)
+	specProvider := mock.NewSpecProvider()
+	chainTime, err := standardchaintime.New(ctx,
+		standardchaintime.WithLogLevel(zerolog.Disabled),
+		standardchaintime.WithGenesisProvider(genesisProvider),
+		standardchaintime.WithSpecProvider(specProvider),
+	)
+	require.NoError(t, err)
+
+	nullSubmitter, err := nullsubmitter.New(ctx)
+	require.NoError(t, err)
+	mockETH2Client, err := mocketh2client.New(ctx)
+	require.NoError(t, err)
+
+	mockValidatingAccountsProvider := mockaccountmanager.NewValidatingAccountsProvider()
+	mockSyncCommitteeAggregator := mocksynccommitteeaggregator.New()
+	mockSigner := mocksigner.New()
+
+	epoch := phase0.Epoch(100)
+	slot := phase0.Slot(epoch * 32)
+
+	params := []standard.Parameter{
+		standard.WithLogLevel(zerolog.Disabled),
+		standard.WithProcessConcurrency(1),
+		standard.WithMonitor(nullmetrics.New(ctx)),
+		standard.WithChainTimeService(chainTime),
+		standard.WithSyncCommitteeAggregator(mockSyncCommitteeAggregator),
+		standard.WithSpecProvider(specProvider),
+		standard.WithBeaconBlockRootProvider(mockETH2Client),
+		standard.WithSyncCommitteeMessagesSubmitter(nullSubmitter),
+		standard.WithValidatingAccountsProvider(mockValidatingAccountsProvider),
+		standard.WithSyncCommitteeRootSigner(mockSigner),
+		standard.WithSyncCommitteeSelectionSigner(mockSigner),
+		standard.WithSyncCommitteeSubscriptionsSubmitter(nullSubmitter),
+	}
+	tests := []struct {
+		name                       string
+		params                     []standard.Parameter
+		dummyRecordsToCreate       int
+		slotToVerifyBeforeAndAfter phase0.Slot
+		removed                    bool
+	}{
+		{
+			name:                       "SuccessfulCleanUp",
+			params:                     params,
+			dummyRecordsToCreate:       110,
+			slotToVerifyBeforeAndAfter: phase0.Slot(50),
+			removed:                    true,
+		},
+		{
+			name:                       "CleanUpSkipped",
+			params:                     params,
+			dummyRecordsToCreate:       50,
+			slotToVerifyBeforeAndAfter: phase0.Slot(32),
+			removed:                    false,
+		},
+	}
+
+	for _, test := range tests {
+		t.Run(test.name, func(t *testing.T) {
+			messageService, err := standard.New(ctx, test.params...)
+			require.NoError(t, err)
+
+			// Create historic records.
+			for i := range test.dummyRecordsToCreate {
+				historicSlot := slot - phase0.Slot(i)
+				messageService.UpdateSyncCommitteeDataRecord(historicSlot, phase0.Root{}, nil)
+			}
+			if test.slotToVerifyBeforeAndAfter > 0 {
+				_, found := messageService.GetDataUsedForSlot(slot - test.slotToVerifyBeforeAndAfter)
+				require.Equal(t, true, found)
+			}
+
+			// Call the clean up method.
+			messageService.RemoveHistoricDataUsedForSlotVerification(slot)
+
+			// Assert we have removed the data.
+			if test.slotToVerifyBeforeAndAfter > 0 {
+				_, found := messageService.GetDataUsedForSlot(slot - test.slotToVerifyBeforeAndAfter)
+				require.Equal(t, !test.removed, found)
 			}
 		})
 	}